Add files via upload
This code trains a Support Vector Machine (SVM) model using the PIMA Diabetes dataset to predict if a person is diabetic or not. The data is first standardized, split into training and test sets, and then used to train the SVM model with an RBF kernel. After training, the model's accuracy is evaluated on both training and test data, and a predictive system is implemented to classify new input data as diabetic or non-diabetic.